feature(lodestar): Missions.Hazard becomes HazardLevel, a named variant of INT64, and Assessor Asa, granted the hazard only while a mission is open, walks the sector's missions by a hazard her grid does not display, the closed missions through the NULL region in Spanner's placement
HazardLevel is declared beside the Mission struct as the named-variant column
the paging demonstration rides. The validator's bounds, the create default,
and every grant's vocabulary compile unchanged: the attribute keeps the number
type through the underlying INT64, the generated TypeScript still types the
field as a number, and the briefing's fold converts explicitly.

The new domain role Assessor holds one List grant on Missions: the title
unconditionally and the hazard under state = 'open', so the CASE stays in the
query and the copy is the visible projection. Persona assessor (Assessor Asa,
Hazard board, Anvil only) joins users.json, the login manifest, the README's
persona table, and the walkthrough, where one check pages her hazard sort two
rows at a time over every Anvil mission, rows carrying id and title alone.

TestPaging_namedVariantKey walks missions?columns=id,title&sort=hazard&limit=2
as the assessor, ascending and descending, forward by next and back by prev:
the sixteen missions no longer open walk the NULL region first ascending and
last descending, the open missions follow by hazard with the primary key
breaking ties, every row is seen exactly once in each direction, and the rows
carry exactly the columns asked for. The two-way walk and the projected-keys
reader are shared with TestPaging_unselectedSortKey. The deploy raises no new
warning: the hazard is neither indexed nor a filter key. Registry key
paging.named-variant-key; DEMONSTRATIONS.md regenerated.

+++ resource/lodestar/pkg/rpc/compile_briefing.go
@@ -100,7 +100,7 @@ func (m *CompileBriefing) Execute(ctx context.Context, client resource.Client, _
 			return nil, errors.Wrap(err, "resources.MissionQuery.List()")
 		}
 		briefing.Missions++
-		briefing.WorstHazard = max(briefing.WorstHazard, row.Data.Hazard)
+		briefing.WorstHazard = max(briefing.WorstHazard, int64(row.Data.Hazard))
 		if row.Masked("fee") {
 			briefing.FeesRedacted++
 		} else {
